chore: Fix API index generation by ignoring "compute small"
The google/cloud/compute/v1 directory in googleapis contains two copies of the API: compute and compute_small.
This is unconventional, and causes the API index generator to fail.

For the moment, we'll just explicitly ignore compute_small. I'll work with the team to try to get it removed from googleapis, so we can revert this commit.