chore: Deprecate host_name, add name_in_service_config
This should be backward compatible assuming all consumers are happy
to ignore new fields. (That's the recommended approach, and it's
been fine in the past.) It may generate warnings for some consumers;
if so they can either ignore the warning or move to use the new
field name.

While `name_in_service_config` is undeniably verbose, it accurate
reflects the nature of the value without trying to describe it
(ambiguously) as the "service name".